Use lazy matching where possible. Add the miq_message_latency field
Latency is the time between when the message was created or when it was marked for delivery (deliver on). When no deliver on date/time is set, latency = dequeued in time. When deliver on was set, latency is the difference between the timestamp of the message and the deliver on datetime